♻ Refactored connectorPlugin class
